requirements.txt
discord.py>=2.5.2
pillow>=11.3.0
python-dotenv>=1.1.1